moves styles to separate directory
Also makes a couple small tweaks to ConfDialog:
  - more accurate text
  - can hit enter instead of clicking on button
