+++ src/styles/ConfDialog.css
